remove main from badaas and create badgen
