--- commands/init.go
@@ -1,27 +1,0 @@
-package commands
-
-import (
-	"strings"
-
-	"github.com/ditrit/badaas/configuration"
-	"github.com/ditrit/badaas/services/userservice"
-	"go.uber.org/zap"
-)
-
-// Create a super user
-func createSuperUser(
-	config configuration.InitializationConfiguration,
-	logger *zap.Logger,
-	userService userservice.UserService,
-) error {
-	// Create a super admin user and exit with code 1 on error
-	_, err := userService.NewUser("admin", "admin-no-reply@badaas.com", config.GetAdminPassword())
-	if err != nil {
-		if !strings.Contains(err.Error(), "already exist in database") {
-			logger.Sugar().Errorf("failed to save the super admin %w", err)
-			return err
-		}
-		logger.Sugar().Infof("The superadmin user already exists in database")
-	}
-	return nil
-}
