--- docker/api/badaas.yml
@@ -1,7 +1,0 @@
-database:
-  host: badaas-db
-  port: 26257
-  sslmode: disable
-  username: root
-  password: postgres
-  name: badaas_db
