--- docker/postgresql/docker-compose.yml
@@ -1,16 +1,0 @@
-version: '3.5'
-
-services:
-  postgres:
-    container_name: postgres_db
-    image: postgres:latest
-    environment:
-      POSTGRES_USER: ${POSTGRES_USER:-postgres}
-      POSTGRES_PASSWORD: ${POSTGRES_PASSWORD:-postgres}
-      POSTGRES_DB: badaas_db
-      PGDATA: /data/postgres
-    volumes:
-      - .:/docker-entrypoint-initdb.d/
-    ports:
-      - "5432:5432"
-    restart: unless-stopped
