--- docker/postgresql/init.sql
@@ -1,1 +1,0 @@
-CREATE EXTENSION "uuid-ossp";
