--- resources/api.go
@@ -1,4 +1,0 @@
-package resources
-
-// Version of Badaas
-const Version = "UNRELEASED"
