--- router/middlewares/JSONHandler.go
@@ -1,10 +1,0 @@
-package middlewares
-
-import (
-	"net/http"
-
-	"github.com/ditrit/badaas/httperrors"
-)
-
-// This handler return a Marshable object and/or an [github.com/ditrit/badaas/httperrors.HTTPError]
-type JSONHandler func(w http.ResponseWriter, r *http.Request) (any, httperrors.HTTPError)
