+++ router/middlewares/middlewareLogger.go
@@ -5,10 +5,15 @@ import (
 	"net/http"
 
 	"github.com/ditrit/badaas/configuration"
+	"github.com/gorilla/mux"
 	"github.com/noirbizarre/gonja"
 	"github.com/noirbizarre/gonja/exec"
 	"go.uber.org/zap"
 )
+
+func AddLoggerMiddleware(router *mux.Router, middlewareLogger MiddlewareLogger) {
+	router.Use(middlewareLogger.Handle)
+}
 
 // Log the requests data
 type MiddlewareLogger interface {
