--- router/router_test.go
@@ -1,24 +1,0 @@
-package router
-
-import (
-	"net/http"
-	"testing"
-
-	controllersMocks "github.com/ditrit/badaas/mocks/controllers"
-	middlewaresMocks "github.com/ditrit/badaas/mocks/router/middlewares"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/mock"
-)
-
-func TestSetupRouter(t *testing.T) {
-	jsonController := middlewaresMocks.NewJSONController(t)
-	middlewareLogger := middlewaresMocks.NewMiddlewareLogger(t)
-	authenticationMiddleware := middlewaresMocks.NewAuthenticationMiddleware(t)
-
-	basicController := controllersMocks.NewBasicAuthenticationController(t)
-	informationController := controllersMocks.NewInformationController(t)
-	eavController := controllersMocks.NewEAVController(t)
-	jsonController.On("Wrap", mock.Anything).Return(func(response http.ResponseWriter, request *http.Request) {})
-	router := SetupRouter(jsonController, middlewareLogger, authenticationMiddleware, basicController, informationController, eavController)
-	assert.NotNil(t, router)
-}
