+++ tools/badgen/Makefile
@@ -1,0 +1,5 @@
+build:
+	go build .
+
+link:
+	ln -sf /home/user/ditrit/badaas/badaas/tools/badgen/badgen /home/user/go/bin/badgen
