+++ tools/badgen/docker/.dockerignore
@@ -1,0 +1,1 @@
+badaas/
