+++ tools/badgen/docker/api/badaas.yml
@@ -1,0 +1,8 @@
+database:
+  host: badaas-db
+  # TODO must be modified according to the db_provider
+  port: 26257
+  sslmode: disable
+  username: root
+  password: postgres
+  name: badaas_db
