+++ tools/badgen/docker/api/docker-compose.yml
@@ -1,0 +1,13 @@
+version: '3.5'
+
+services:
+  api:
+    build:
+      # TODO
+      context: ../../../..
+      dockerfile: ./example/badaas/docker/api/Dockerfile
+    container_name: "badaas-api"
+    ports:
+      - "8000:8000"
+    depends_on:
+      - db
