+++ tools/badgen/docker/cockroachdb/docker-compose.yml
@@ -1,0 +1,15 @@
+version: '3.5'
+
+services:
+  db:
+    image: cockroachdb/cockroach:latest
+    container_name: badaas-db
+    ports:
+      - "26257:26257"
+      - "8080:8080" # Web based dashboard
+    command: start-single-node --insecure
+    volumes:
+      - "${PWD}/badaas/.cockroach-data:/cockroach/cockroach-data"
+    environment:
+      - COCKROACH_USER=root
+      - COCKROACH_DATABASE=badaas_db
