+++ tools/badgen/docker/postgres/docker-compose.yml
@@ -1,0 +1,16 @@
+version: '3.5'
+
+services:
+  db:
+    image: postgres:latest
+    container_name: badaas-db
+    environment:
+      POSTGRES_USER: ${POSTGRES_USER:-postgres}
+      POSTGRES_PASSWORD: ${POSTGRES_PASSWORD:-postgres}
+      POSTGRES_DB: badaas_db
+      PGDATA: /data/postgres
+    volumes:
+      - .:/docker-entrypoint-initdb.d/
+    ports:
+      - "5432:5432"
+    restart: unless-stopped
