+++ tools/badgen/docker/postgres/init.sql
@@ -1,0 +1,1 @@
+CREATE EXTENSION "uuid-ossp";
