+++ tools/badgen/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/ditrit/badaas/tools/badgen/cmd"
+
+func main() {
+	cmd.Execute()
+}
