+++ tools/badgen/scripts/run.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+
+docker compose -f "badaas/docker/db/docker-compose.yml" -f "badaas/docker/api/docker-compose.yml" up --build -d
